Re-implemented integration using Cohen, N.S.V. 2013, 'Open Fusion', pp. 1–35.
Updated the integration function to accept a camera rather than a raw pose and moved many of the
coordinate system conversions into the camera since this is now possible and a logical place for them to be.

Future changes may need to inline these to imrpove perfomance as there are a number of areas where we
calculate things which we could have cached and where we call a function so many times that the overhead becomes
significant.