Implemented move to and look at methods for Camera with tests
These convenience methods allow for the camera pose to be modified in a more intuitive way rather than
forcing the client to generate the pose matrix.

Move To changes the location of the camera in global coords

Look At changes the rotation of the camera such that it faces a given glbal coordinate. It's based on the
gluLookAt code however it generates the inverse matrix since it's the camera that is being moved rather
than the world coordinate frame