Simplified Camera Coordinate Methods
Camera coordinate conversion methods now return their value rather than taking it as a reference parameter
as this makes the code easier to read.

In the process, discovered that the raycast method was constructing vertices and normals in cam space where global space
is required. Also the rendering helper function was both
- assuming that coordinats where in camera space and
- miscalculating the shading. This is updated