Modified integrate to reflect the Newcombe approach of comparing depths
Originally I compared distance using the approach from Cohen et al.  They'd modified the p-code
in Izadi which comapres depth to distance (incorrectly).

I've now reverted to the Newcombe method which compares depths not distances.  This also means the TSDF should
contain values in the range -1 to 1 and will require changes to the raycasting code to bring the two into line