Modified Volume::point_to_voxel to return a voxel rather than taking one as a reference
Brings this method into line with other coordinate conversion routines.

Also removed the code which constrained the value to be in the grid. Now a flag is populated to determine
whether the provided voxel is in the grid or not. It's up to the calling code to determine if this is OK
or not.

Modified calling code to correctly manage this throughout