Fixed numerical stability issue with raycaster
When multiplying the trilinearly interpolated TSDF by the step size we get a number which is so small that
it effectively doesn;t change the t value and we end up in an infinite loop. By making sure we always advance at least
1mm we avoid this though the render takes longer. We may want to tune this number in future.

Modified the test harness to generate 512x512 image for comparison.