Handle camera in Volume
Previoulsy when finding the start of the path to walk for a ray, the case where the camera was in the Volume was
not handled. This leads to increased processing time (walking voxels behind the vamera) and inaccurate rendering
at higher resolutions where intersections were recorded as being behind the camra - this may be indicative of a
probblem with integration too.

Now a test is performed to see if the origin is inside the voxel grid and if so, t is set to 0 and the start point
is set to the camera origin

Also put tighter checking around memory allocation, removing asserts and replacing with exceptions