Introduced Factory and refactored structue
Hid Raycaster because it's not like we can have an arbitrary raycaster operate on an arbitrary TSDFVolume. Raycaster is explicitly a CPU or GPU
class and will be an implementation detail of the TSDFVolume so we've returned to having a specific method call for raycasting on the volume.

Created a factory method in TSDFVolume to instantiate a specific subclass. Updated main.cpp to use the factory for test purposes

Created a PngWrapper class to wrap up width, height and data.  This saves every class from having to know about load/save PNG.

Removed some unused files from the porject

Restructured the directory structure to separate CPU and GPU implementations whil ekeeping common code common.