fix: Remove data access functions from TSDFVolume public interface
In the past these made sense to surface however in the new model, all
the data being accessed exists on the GPU Device and so having them
in the public interface is dangerous as any class could call them
leading to a memory error.

They've now been made internal to the class and use the forceinline
attribute and device attrbutes to ensure that they run at reasonable
speed.

A further optimisation would be to just replace the inline occurrences
and we may still do that.