feature: Update integration code to use deformation field
The old integration code projected the voxel centres into the depth map
These voxel centres were computed assuming that the voxel grid did not
deform which was OK at that time.

Now the voxel grid can defom and we cannot compute the voxel centre so
the code is updated to use the voxel transformations as centres.