refactor: Removing CPU based TSDF and related Eigen dependency
The abstract TSDF class defined the interface using Eigen types eg Vector3i.

The internals of the GPU version uses CUDA vector types (dim3).

This patch replaces the public Eigen based interface with one based on types that
are essentially redefines of the CUDA classes. This preserves the CUD types
only in implemnetation code but removes a dependency on Eigen for the interface