test: Increase threads to see if we can tune performance
Still performs worse than the 'old' code where everything wa integrated.

Not sure why we can't run 1024 threads across the new integrate kernel. It's marginally larger
than the old one in that it dreferences the data index but no reason why it should be so much heavier
on registers that  can see.