Test: Remove the max weight cap
Checking to see if the weight cap makes a difference in umbrella reconstruction.
We had an issue where up to fram 100, all was wllell with reocnstruction but beyond that we get a deviation where the canonoical frame starts to track the current depth frame.

Removing the weight cap didn't fix it.