Refactor: Modify MarkAndSweepMC to flag impacted voxels
This refactoring removes the former data collection fduring marching cubes which would produce a list of impacted voxel indices along wit the number
of updates made to them.  The new approach flags all impacted voxels in a single flag array. No count is kept. the aim is to use this to
track impacted voxels and to generate the energy equation used to update the deformation field

This change als added cudaSafeCopy... utility methods